requirements.txt
asyncio==4.0.0
importlib==1.0.4
MouseInfo==0.1.3
pathlib==1.0.1
pillow==12.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
PyYAML==6.0.3
uv==0.9.3
websockets==15.0.1
